Terminado mapa e lista
- TODO: Pedido de permissões;
- TODO: Tela de previsão do tempo;

+++ app/src/main/java/com/devmaker/luizmatias/weatherapp_devmaker/Clima.java
@@ -7,11 +7,11 @@ package com.devmaker.luizmatias.weatherapp_devmaker;
 public class Clima {
 
     private String descricao; //descrição do clima
-    private float temperatura; //temperatura em celsius
-    private float pressao; //pressão em hPa
-    private float umidade; //porcentagem de umidade
-    private float temperatura_minima; //temperatura minima em celsius
-    private float temperatura_maxima; //temperatura maxima em celsius
+    private double temperatura; //temperatura em celsius
+    private double pressao; //pressão em hPa
+    private double umidade; //porcentagem de umidade
+    private double temperatura_minima; //temperatura minima em celsius
+    private double temperatura_maxima; //temperatura maxima em celsius
 
     public String getDescricao() {
         return descricao;
@@ -21,43 +21,43 @@ public class Clima {
         this.descricao = descricao;
     }
 
-    public float getTemperatura() {
+    public double getTemperatura() {
         return temperatura;
     }
 
-    public void setTemperatura(float temperatura) {
+    public void setTemperatura(double temperatura) {
         this.temperatura = temperatura;
     }
 
-    public float getPressao() {
+    public double getPressao() {
         return pressao;
     }
 
-    public void setPressao(float pressao) {
+    public void setPressao(double pressao) {
         this.pressao = pressao;
     }
 
-    public float getUmidade() {
+    public double getUmidade() {
         return umidade;
     }
 
-    public void setUmidade(float umidade) {
+    public void setUmidade(double umidade) {
         this.umidade = umidade;
     }
 
-    public float getTemperatura_minima() {
+    public double getTemperatura_minima() {
         return temperatura_minima;
     }
 
-    public void setTemperatura_minima(float temperatura_minima) {
+    public void setTemperatura_minima(double temperatura_minima) {
         this.temperatura_minima = temperatura_minima;
     }
 
-    public float getTemperatura_maxima() {
+    public double getTemperatura_maxima() {
         return temperatura_maxima;
     }
 
-    public void setTemperatura_maxima(float temperatura_maxima) {
+    public void setTemperatura_maxima(double temperatura_maxima) {
         this.temperatura_maxima = temperatura_maxima;
     }
 }
